requirements.txt
flask
flask-socketio
eventlet
pyjwt
flask-session
mysql
mysql-connector

